requirements.txt
asgiref==3.4.1
asn1crypto==0.24.0
beautifulsoup4==4.7.1
Brlapi==0.6.7
certifi==2018.8.24
chardet==3.0.4
chrome-gnome-shell==0.0.0
cryptography==2.6.1
cupshelpers==1.0
distro==1.3.0
distro-info==0.21
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-extensions==3.1.3
django-six==1.0.4
entrypoints==0.3
feedparser==5.2.1
galternatives==1.0.4
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
keyring==17.1.1
keyrings.alt==3.1.1
louis==3.8.0
lxml==4.3.2
Mako==1.0.7
MarkupSafe==1.1.0
meteo-qt==1.0.0
musicbrainzngs==0.6
mutagen==1.40.0
olefile==0.46
pexpect==4.6.0
Pillow==5.4.1
pipenv==2021.5.29
psutil==5.5.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
PyGObject==3.30.4
pyinotify==0.9.6
PySimpleSOAP==1.16.2
pysmbc==1.0.15.6
python-apt==1.8.4.3
python-debian==0.1.35
python-debianbts==2.8.2
python-pam==1.8.4
python-xapp==1.2.0
python-xlib==0.23
pytz==2021.3
pyxattr==0.6.1
pyxdg==0.25
quodlibet==4.2.1
reportbug===7.5.3-deb10u1
reportlab==3.5.13
requests==2.21.0
scour==0.37
screen==1.0.1
SecretStorage==2.3.1
setproctitle==1.1.10
six==1.12.0
soupsieve==1.8
sqlparse==0.4.2
typing-extensions==3.10.0.2
unattended-upgrades==0.1
urllib3==1.24.1
virtualenv==15.1.0
virtualenv-clone==0.5.7
webencodings==0.5.1
xlwt==1.3.0
youtube-dl==2019.1.17
